Refactor main visualise_tcav.py code into visualise.py
